Fix: Updated unit tests
